+++ front-end/Classes/Sprite/HealPack.h
@@ -9,12 +9,12 @@ class HealPack : public cocos2d::Sprite {
 	HealPack();
 
 	static std::map<std::string, HealPack*> pool;
-	float hp;
+	int hp;
 	std::string id;
 public:
 	static HealPack* create(rapidjson::GenericValue<UTF8<>> &data);
 	static void remove(const std::string &id);
 	void broadcastEaten() const;
-	float getHp() const;
+	int getHp() const;
 	const std::string& getId() const;
 };
